moved App container to separate folder
