+++ src/containers/App/index.tsx
@@ -1,10 +1,10 @@
 import * as React from 'react';
 import "tachyons";
-import './App.css';
-import Logo from "./components/Logo";
-import Navigation from "./components/Navigation";
-import Rank from "./components/Rank";
-import ImageLinkForm from "./components/ImageLinkForm";
+import './index.css';
+import Logo from "../../components/Logo";
+import Navigation from "../../components/Navigation";
+import Rank from "../../components/Rank";
+import ImageLinkForm from "../../components/ImageLinkForm";
 import Particles from 'react-particles-js';
 
 const particlesOptions = {
