+++ src/index.tsx
@@ -1,8 +1,8 @@
-import * as React from 'react';
-import * as ReactDOM from 'react-dom';
-import App from './App';
-import './index.css';
-import registerServiceWorker from './registerServiceWorker';
+import * as React from "react";
+import * as ReactDOM from "react-dom";
+import App from "./containers/App";
+import "./index.css";
+import registerServiceWorker from "./registerServiceWorker";
 
 ReactDOM.render(
   <App />,
